Engine improvements
Vector3D renamed to Point3D as it only contains a position and operation overloads. New Vector3D struct instead takes in the product of two points (a tail point subtracting the head point) and can be used to calculate actual vectors for objects.

Do Note: While Points may only represent XYZ, they can still be used for things like Movement. Vector3D is there to help differentiate what is a true point, and what is an actual "direction".

For now, quad planes are barely functional - it's a test bed for points and vectors as of right now.