Model Testing
Created a Model Structure that holds a vector of Vertices and a vector of Elements, with Vertices dictating point position and elements dictating the actual connections between them. The Model also contains an origin point from which all vertices will be transformed by.

The Model is currently hard coded to a basic 3D cube, with both vertex placement and element ordering.

Honestly, running into difficulties progressing due to lack of understanding on how 3D models and the pipeline works on a very deep technical level. While I understand principles such as buffers (Vertex, Element, etc) I am struggling to apply it in a non-bugged manner. I also do not know how to parse information from a text file (such as vertices placement) into the program as readable data.

Due to this, the OpenGL branch of this project will likely be developed on instead, as OpenGL handles all of the difficulties that I am facing. Additionally, the 3D engine is merely a means to an end to displaying 3D movement and positioning.