Further work on OpenGL
Moved the Graphics into its own class, repeated the same structure of Physics (LINK)-> Graphics in the refactor branch, where Physics dictates the simulation loop while refreshing the window.

Shader classes developed, now it's a matter of developing a way to load vertex or fragment shader code into the program. Ifstreams seem to be the way to go.